tests/wpt/web-platform-tests/direct-sockets/open-securecontext.http.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Sockets test: Open from non-secure context</title>
    <link rel="help" href="https://github.com/WICG/direct-sockets/blob/main/docs/explainer.md#security-considerations">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
        test(() => {
          assert_false(typeof TCPSocket !== 'undefined', 'TCPSocket is present');
        }, 'TCPSocket must be undefined in non-secure context');

        test(() => {
          assert_false(typeof UDPSocket !== 'undefined', 'UDPSocket is present');
        }, 'UDPSocket must be undefined in non-secure context');
    </script>
  </body>
</html>
